+++ docker-compose/restart.sh
@@ -1,2 +1,2 @@
 docker compose down
-docker compose up
+docker compose up -d